deps(pinyin): move to only be in tui
As that is the only place that uses it
