+++ lib/Cargo.toml
@@ -41,7 +41,6 @@ md5.workspace = true #  = "0.7"
 num-bigint.workspace = true #  = "0.4"
 opml.workspace = true #   = "1"
 parking_lot.workspace = true
-pinyin.workspace = true #  = "0.9"
 quick-xml.workspace = true #  = "0.28"
 rand.workspace = true #  = "0.8"
 regex.workspace = true #  = "^1.5.5"
