+++ lib/src/utils.rs
@@ -4,29 +4,10 @@ use std::iter::FusedIterator;
 use std::path::{Path, PathBuf};
 
 use anyhow::{Context, Result, anyhow};
-use pinyin::ToPinyin;
 use rand::RngExt;
 use unicode_segmentation::UnicodeSegmentation;
 
 use crate::config::ServerOverlay;
-
-#[must_use]
-pub fn get_pin_yin(input: &str) -> String {
-    let mut b = String::new();
-    for (index, f) in input.to_pinyin().enumerate() {
-        match f {
-            Some(p) => {
-                b.push_str(p.plain());
-            }
-            None => {
-                if let Some(c) = input.to_uppercase().chars().nth(index) {
-                    b.push(c);
-                }
-            }
-        }
-    }
-    b
-}
 
 // TODO: decide filetype supported by backend instead of in library
 #[must_use]
@@ -339,14 +320,6 @@ mod tests {
 
     use super::*;
     use pretty_assertions::assert_eq;
-
-    #[test]
-    fn test_pin_yin() {
-        assert_eq!(get_pin_yin("陈一发儿"), "chenyifaer".to_string());
-        assert_eq!(get_pin_yin("Gala乐队"), "GALAledui".to_string());
-        assert_eq!(get_pin_yin("乐队Gala乐队"), "leduiGALAledui".to_string());
-        assert_eq!(get_pin_yin("Annett Louisan"), "ANNETT LOUISAN".to_string());
-    }
 
     #[test]
     fn test_substr() {
