+++ tui/Cargo.toml
@@ -64,6 +64,7 @@ pathdiff.workspace = true
 either.workspace = true
 lru.workspace = true
 mime_guess.workspace = true
+pinyin.workspace = true
 
 [target.'cfg(windows)'.dependencies]
 # required for setting "DETACHED_PROCESS" for spawning the server
