+++ tui/src/ui/utils.rs
@@ -1,3 +1,4 @@
+use pinyin::ToPinyin;
 use tuirealm::{
     props::{Style, TextModifiers},
     ratatui::layout::{Constraint, Layout, Rect},
@@ -5,6 +6,24 @@ use tuirealm::{
 
 /// Save a few characters and lines by having a alias.
 pub const STYLE_REMOVE_REVERSE: Style = Style::new().remove_modifier(TextModifiers::REVERSED);
+
+#[must_use]
+pub fn get_pin_yin(input: &str) -> String {
+    let mut b = String::new();
+    for (index, f) in input.to_pinyin().enumerate() {
+        match f {
+            Some(p) => {
+                b.push_str(p.plain());
+            }
+            None => {
+                if let Some(c) = input.to_uppercase().chars().nth(index) {
+                    b.push(c);
+                }
+            }
+        }
+    }
+    b
+}
 
 // /// Get block
 // pub fn get_block<'a>(props: &Borders, title: (String, Alignment), focus: bool) -> Block<'a> {
@@ -70,6 +89,14 @@ pub fn draw_area_top_right_absolute(parent: Rect, width: u16, height: u16) -> Re
 mod tests {
     use super::*;
     use pretty_assertions::assert_eq;
+
+    #[test]
+    fn test_pin_yin() {
+        assert_eq!(get_pin_yin("陈一发儿"), "chenyifaer".to_string());
+        assert_eq!(get_pin_yin("Gala乐队"), "GALAledui".to_string());
+        assert_eq!(get_pin_yin("乐队Gala乐队"), "leduiGALAledui".to_string());
+        assert_eq!(get_pin_yin("Annett Louisan"), "ANNETT LOUISAN".to_string());
+    }
 
     #[test]
     fn test_utils_ui_draw_area_in() {
